refactor: app constants
